int to int64

+++ internal/repository/fibonacci.go
@@ -15,8 +15,8 @@ func NewFibonacciRepository(rdbCache *cache.Cache) *FibonacciRepository {
 	return &FibonacciRepository{rdbCache: rdbCache, key: "fibonacci"}
 }
 
-func (r *FibonacciRepository) GetCachedFibSequence(ctx context.Context, end int) ([]int, error) {
-	var result []int
+func (r *FibonacciRepository) GetCachedFibSequence(ctx context.Context, end int) ([]int64, error) {
+	var result []int64
 
 	if err := r.rdbCache.Get(ctx, fmt.Sprintf("%s:%d", r.key, end), &result); err != nil {
 		return nil, err
@@ -25,7 +25,7 @@ func (r *FibonacciRepository) GetCachedFibSequence(ctx context.Context, end int)
 	return result, nil
 }
 
-func (r *FibonacciRepository) SetFibSequence(ctx context.Context, fibSequence []int, end int) error {
+func (r *FibonacciRepository) SetFibSequence(ctx context.Context, fibSequence []int64, end int) error {
 	toSet := cache.Item{
 		Ctx:   ctx,
 		Key:   fmt.Sprintf("%s:%d", r.key, end),
